requirements.txt
Flask==2.1.1
Werkzeug==2.1.1
torch==2.0.0
torchaudio==2.0.1
transformers==4.31.0
requests==2.28.1
ffmpeg-python==0.2.0
flask-socketio==5.3.1
numpy==1.23.5
scipy==1.9.1
tensorflow-intel==2.16.1
daal==2021.4.0